add audio_slice
An audio_slice is either a constant or a span. This concept
allows now to split the amplify_processor into a smoother_processor and
multiply_processor. smoother_processor receives an event an generates a
smoothed ramp from previous value to new value. multiply_processor can
take this ramp and a signal and realize our amplification therefore.