Add batch_middleware
The batch_middleware waits for a batch_action to occur. If a batch_action
occurs, it will execute all batched actions in sequence. It will set a referenced
flag, that batching is currently in progress. This flag can be used to postpone
store subscription updates, until the last action in the batch is executed.