Solo rework
Introduce solo groups. A solo group is a set of mixer channels with
the same mix target. Soloing a channel in the group will mute other
channels in the group. Exclusive solo is removed for now, as it
doesn't really work with MIDI assigned controls. Maybe in the future
we will readd this feature.