FxChain: Increase cacheBuffer to avoid delegate destruction
FxChain delegate creation is visible on raspi2. Therefore
when scrolling through chain you always see a busy loader,
trying to recreate new delegates which become visible. To
avoid the destruction of delegates which become invisible
and therefore to avoid the recreation of those, we increase
the cacheBuffer property. This way we can keep quite a few
delegates alive during scrolling.