runtime: add delete_channel_address
Previously if a channel was deleted and it was used somewhere as an input,
then this input was reset to default state, the 'Mix' state. This might
cause a feedback loop in some routing situations. Therefore a new address
state is introduced for deleted channels. The user then needs to set the
input manually to some device/channel or leave it in this state.